Update for PowerTimer
I add a new PowerTimerEntry screen, this in the OE-A is linked to the TimerEntry screen, I don't like it because every time a new TimerEntry recording timer is added, you don't need the PowerTimerEntry self code "ok" and "cancel" pixmap. as a consequence there is always an "no found" widget ok and cancel in skin. This differentiates them.